+++ config/application.rb
@@ -25,6 +25,7 @@ module App
     end
 
     config.active_job.queue_adapter = :sidekiq
-    config.paths.add "packs", glob: "*/{*,*/concerns}", eager_load: true
+    # config.paths.add "packs", glob: "oauth/app/*/**", eager_load: true
+    config.autoload_paths << "#{root}/packs/oauth/app"
   end
 end
