+++ config/routes.rb
@@ -8,7 +8,7 @@ Rails.application.routes.draw do
   mount Sidekiq::Web => "/sidekiq"
   mount Rswag::Ui::Engine => "/api-docs"
   mount Rswag::Api::Engine => "/api-docs"
-  devise_for :users, controllers: { omniauth_callbacks: "oauth/omniauth_callbacks" }
+  devise_for :users, controllers: { omniauth_callbacks: "oauth/controllers/omniauth_callbacks" }
   # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html
 
   # Defines the root path route ("/")
